added log_info, log_warn and log_err utility functions; fixed cleanup function

+++ backup.sh
@@ -19,8 +19,12 @@ REPOSITORY="$(realpath --relative-to "$SCRIPT_DIR" "$REPOSITORY")"
 
 # ---- UTILS ----
 
+log_err () { echo "ERROR: $1" >&2; }
+log_warn () { echo "WARN: $1" >&2; }
+log_info () { echo "$1"; }
+
 send_telegram () {
-    echo -e "sending Telegram message"
+    log_info -e "sending Telegram message"
     curl -X POST \
         --no-progress-meter \
         -H 'Content-Type: application/json' \
@@ -48,34 +52,35 @@ send_succ () {
 
 exit_handler () {
     retval=$?
+
     # define handler for when cleanup fails
     # note that it is not possible trap EXIT inside of exit handler
     err_handler () {
-        echo "ERROR: backup environment cleanup failed"
+        log_err "backup environment cleanup failed"
         send_error "backup environment cleanup failed"
         exit 1
     }
     trap err_handler ERR
 
-    echo ""
+    log_info ""
     if [[ ${IS_LAUNCHED+1} ]]; then
         if ((retval==255)); then
-            echo "ERROR: error occurred or received exit signal"
+            log_err "error occurred or received exit signal"
             send_error "error occurred or received exit signal"
         elif ((retval==0)); then
-            echo "backup completed"
+            log_info "backup completed"
             send_succ "backup"
         else
-            echo "ERROR: backup failed with code $retval"
+            log_err "backup failed with code $retval"
             send_error "backup failed with code $retval"
         fi
     else
         if ((retval==0)); then
             # this means that exit 0 was called before the backup was started
-            echo "finished without backing up"
+            log_info "finished without backing up"
             send_succ "finished without backing up"
         else
-            echo "ERROR: backup was not started: an error occurred"
+            log_err "backup was not started: an error occurred"
             send_error "backup was not started: an error occurred"
         fi
     fi
@@ -85,12 +90,10 @@ exit_handler () {
 }
 
 cleanup () {
-    echo "removing temporary backup environment"
-    if [[ ${MOUNT_PID+1} ]]; then
-        rm -d "$MOUNT_PATH"
-        rm -d "$ROOT/$SOURCE_NAME"
-        rm -rf "$ROOT"
-    fi
+    log_info "removing temporary backup environment"
+    if [[ -d $ROOT/$SOURCE_NAME ]]; then rm -d "$ROOT/$SOURCE_NAME"; fi
+    if [[ -d $ROOT/repository ]]; then rm -d "$ROOT/repository"; fi
+    rm -rf "$ROOT"
 }
 
 trap exit_handler EXIT
@@ -99,23 +102,23 @@ trap exit_handler EXIT
 
 
 # construct fake restic root dir
-echo "creating temporary backup environment"
+log_info "creating temporary backup environment"
 ROOT="$(mktemp -d)"
 mkdir "$ROOT"/{tmp,repository,"$SOURCE_NAME"}
 
 # copy restic executable
 restic_path="$(which restic)" && true
-if (($?>0)); then echo "ERROR: could not locate restic"; exit 1; fi
+if (($?>0)); then log_err "could not locate restic"; exit 1; fi
 cp "$(realpath "$restic_path")" "$ROOT/restic"
 
-
-# construct and run the backup command
+# construct command
 cmd_1=(/restic -r /repository backup --ignore-inode /"$SOURCE_NAME")
 cmd_2=(unshare -rR "$ROOT" "${cmd_1[@]}")
 cmd_3=("$SCRIPT_DIR/serve-rclone-mount.sh" "$SOURCE" --mountpoint "$ROOT/$SOURCE_NAME" -- "${cmd_2[@]}")
 cmd_4=("$SCRIPT_DIR/serve-bindfs-mount.sh" "$REPOSITORY" --mountpoint "$ROOT/repository" -- "${cmd_3[@]}")
 
-echo ""
+# run the command
+log_info ""
 IS_LAUNCHED=1
 "${cmd_4[@]}" && true
 exit $?
